requirements.txt
streamlit
transformers==4.38.2
torch
openai
python-dotenv
langchain==0.2.0
ctransformers==0.2.27
sentence-transformers==2.2.2
chromadb==0.4.4
